Finish ber error mapping, everything compiles
