+++ src/lib.rs
@@ -1,5 +1,5 @@
 #![doc = include_str!("../README.md")]
-// #![cfg_attr(not(test), no_std)]
+#![cfg_attr(not(test), no_std)]
 extern crate alloc;
 
 #[cfg(test)]
